server/requirements.txt
alembic==1.14.1
amqp==5.3.1
aniso8601==10.0.1
# backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
Faker==35.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==2.1.5
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.42
SQLAlchemy-serializer==1.4.12
tldextract==5.1.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
